avoid error from tl.float16 without triton
